+++ crates/swc/tests/fixture/issues-2xxx/2086/output/index.ts
@@ -1,1 +1,1 @@
-export default class t{#a;constructor(t){this.#a=t}#b(){setTimeout(()=>{this.#a.textContent="TESTED"},1e3)}run(){this.#b()}};
+export default class t{#a;constructor(t){this.#a=t}#b(){setTimeout(()=>{this.#a.textContent="TESTED"},1e3)}run(){this.#b()}}