+++ crates/swc/tests/fixture/issues-2xxx/2717/output/index.ts
@@ -38,4 +38,4 @@ export default function n() {
                 ];
         }
     });
-};
+}