+++ crates/swc/tests/tsc-references/typeFromPropertyAssignment5.2.minified.js
@@ -1,5 +1,5 @@
 //// [a.js]
-export default function MyClass() {};
+export default function MyClass() {}
 MyClass.bar = class {
 }, MyClass.bar;
 //// [b.js]