+++ crates/swc/tests/vercel/full/d3-time-format/1/output/index.js
@@ -223,7 +223,7 @@ export default function v(e) {
             }, t;
         }
     };
-};
+}
 var y = {
     "-": "",
     _: " ",