+++ crates/swc/tests/vercel/loader-only/2/output/index.js
@@ -10,7 +10,7 @@ export default function MyComp() {
     var projectId = projectInfo === null || projectInfo === void 0 ? void 0 : projectInfo.id;
     var ref1 = useProjectBranches(projectId), branches = ref1.data;
     return /*#__PURE__*/ _jsx(_Fragment, {});
-};
+}
 for(var i = 0; i < points.length; i++){
     var point = points[i];
     for(var key_ in keys){