+++ crates/swc/tests/vercel/loader-only/next-33291/2-classic/output/index.js
@@ -240,4 +240,4 @@ export default function Home() {
         },
         __self: this
     })))));
-};
+}