+++ crates/swc_ecma_codegen/tests/fixture/vercel/d3-color/1/output.js
@@ -10,7 +10,7 @@ function cubehelixConvert(o) {
 }
 export default function cubehelix(h, s, l, opacity) {
     return 1 === arguments.length ? cubehelixConvert(h) : new Cubehelix(h, s, l, null == opacity ? 1 : opacity);
-};
+}
 ;
 export function Cubehelix(h, s, l, opacity) {
     this.h = +h, this.s = +s, this.l = +l, this.opacity = +opacity;