+++ crates/swc_ecma_codegen/tests/test262-min/e2470430b235b9bb.module.js
@@ -1,1 +1,1 @@
-export default function a(){};let b;export{b as a};
+export default function a(){}let b;export{b as a};