+++ crates/swc_ecma_codegen/tests/test262/c92dc954a0ce196d.module.js
@@ -1,2 +1,2 @@
-export default function a() {};
+export default function a() {}
 false;