+++ crates/swc_ecma_codegen/tests/test262/e2470430b235b9bb.module.js
@@ -1,3 +1,3 @@
-export default function a() {};
+export default function a() {}
 let b;
 export { b as a };