+++ crates/swc_ecma_minifier/tests/fixture/issues/buble/1/output.mangleOnly.js
@@ -1,3 +1,3 @@
 export default function r() {
     throw new Error(`Failed to recognize value \`${value}\` for property ` + `\`${property}\`.`);
-};
+}