+++ crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/001/output.js
@@ -14,4 +14,4 @@ export default function createWidgetsManager(onWidgetsUpdate) {
         update: scheduleUpdate,
         getWidgets: ()=>widgets
     };
-};
+}