+++ crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/001/output.mangleOnly.js
@@ -26,4 +26,4 @@ export default function e(e) {
             return r;
         }
     };
-};
+}