+++ crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/002/output.js
@@ -279,4 +279,4 @@ export default function createInstantSearchManager({ indexName , initialState ={
             skip = !0;
         }
     };
-};
+}