+++ crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/004/output.js
@@ -337,4 +337,4 @@ export default function createInstantSearchManager(param) {
         clearCache: clearCache,
         skipSearch: skipSearch
     };
-};
+}