+++ crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/004/output.mangleOnly.js
@@ -418,7 +418,7 @@ export default function m(n) {
         clearCache: P,
         skipSearch: b
     };
-};
+}
 function g(t) {
     if (!t) {
         return [];