+++ crates/swc_ecma_minifier/tests/fixture/issues/react/hooks/1/output.js
@@ -5,4 +5,4 @@ import useTeam from "@swr/use-team";
 export default function MyComp() {
     var projectName = useRouter().query.project, projectInfo = useProject(projectName).data;
     return useTeam().teamSlug, useProjectBranches(null == projectInfo ? void 0 : projectInfo.id).data, _jsx(_Fragment, {});
-};
+}