+++ crates/swc_ecma_minifier/tests/full/next-33088/output.js
@@ -12,4 +12,4 @@ import*as t from"react";export default function e(){const[e,i]=t.useState({hits:
 
             </ul>
 
-        </>};
+        </>}