+++ crates/swc_ecma_transforms_base/tests/resolver/export/default/class/decl/scope/output.js
@@ -3,5 +3,5 @@ export default class Test__1 {
         let Test__2 = 2;
         console.log(Test__2);
     }
-};
+}
 Test__1 = 2;