+++ crates/swc_ecma_transforms_react/tests/jsx/fixture/issue-1799/case1/output.mjs
@@ -6,5 +6,5 @@ export default function Foo() {
             await doSomething();
         }
     });
-};
+}
 Foo.displayName = "Foo";