+++ crates/swc_ecma_transforms_react/tests/jsx/fixture/issue-5099/1/output.mjs
@@ -16,4 +16,4 @@ export default function Greet(props) {
             __html: props.data.markup.content
         }
     }));
-};
+}