+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - **(css/minifier)** Fix minification of universal selectors (#5826) ([1803874](https://github.com/swc-project/swc/commit/1803874f1484ab2a3c7b23b5e027e72f99740fba))
 
 
+- **(es/codegen)** Remove redundant `;` after export default decl (#5851) ([2b24642](https://github.com/swc-project/swc/commit/2b2464271aa938968986287fa03e2ef06dcd3fd7))
+
+
 - **(es/minifier)** Don't replace parameters of IIFE if it's recursive (#5852) ([a5f7b69](https://github.com/swc-project/swc/commit/a5f7b6946fcf020b4669ac301b944ba5c0adbbba))